finished algorithm

+++ AAAButterLow.h
@@ -11,15 +11,15 @@
 @interface AAAButterLow : NSObject {
 
 }
-@property (nonatomic) float fr;
-@property (nonatomic) float a0;
-@property (nonatomic) float omegaPrime;
-@property (nonatomic) float a1;
-@property (nonatomic) float a2;
-@property (nonatomic) float b1;
-@property (nonatomic) float b2;
-@property (nonatomic) float c;
+@property (nonatomic) double fr;
+@property (nonatomic) double a0;
+@property (nonatomic) double omegaPrime;
+@property (nonatomic) double a1;
+@property (nonatomic) double a2;
+@property (nonatomic) double b1;
+@property (nonatomic) double b2;
+@property (nonatomic) double c;
 
--(id)initWithSampleFreq:(float)sampleFreq_ CutoffFreq:(float)cuttofff;
--(void) filterArray: (float*)dataArray DataLength:(int)dataLength ResultArray:(float*)resultArray ResultLength:(int)resultLength;
+-(id)initWithSampleFreq:(double)sampleFreq_ CutoffFreq:(double)cuttofff;
+-(void) filterArray: (double*)dataArray DataLength:(int)dataLength ResultArray:(double*)resultArray ResultLength:(int)resultLength;
 @end
